#97 - Fix in items for price in float

+++ app/src/main/java/com/projekt/zespolowy/fun_organizer/eventItems/SingleItemSmallModel.kt
@@ -3,5 +3,5 @@ package com.projekt.zespolowy.fun_organizer.eventItems
 data class SingleItemSmallModel(
         var name: String,
         var description: String,
-        var value: Int
+        var value: Float
 )
